+++ renderer/units/JavaScript Runner/samples/Emoji/Write CLDR Annotations to File.js
@@ -1,8 +1,4 @@
 // Write CLDR Annotations to File
-const fs = require ('fs');
-const path = require ('path');
-const { app } = require ('electron').remote;
 const annotations = require ('./lib/unicode/get-cldr-annotations.js');
-const jsonFile = path.join (app.getPath ('desktop'), 'cldr-annotations.json');
-fs.writeFileSync (jsonFile, $.stringify (annotations, null, 4));
+let jsonFile = $.save ($.stringify (annotations, null, 4), 'cldr-annotations.json');
 $.writeln (`Wrote CLDR annotations to JSON file:\n${jsonFile}`);