+++ renderer/units/JavaScript Runner/samples/Emoji/Write Emoji Groups to File.js
@@ -1,8 +1,4 @@
 // Write Emoji Groups to File
-const fs = require ('fs');
-const path = require ('path');
-const { app } = require ('electron').remote;
 const emojiTestGroups = require ('emoji-test-groups');
-const jsonFile = path.join (app.getPath ('desktop'), 'emoji-groups.json');
-fs.writeFileSync (jsonFile, $.stringify (emojiTestGroups, null, 4));
+let jsonFile = $.save ($.stringify (emojiTestGroups, null, 4), 'emoji-groups.json');
 return `Wrote emoji groups to JSON file:\n${jsonFile}`;