+++ renderer/units/JavaScript Runner/samples/Unihan/Write Unihan Compatibility Variants to File.js
@@ -1,7 +1,4 @@
 // Write Unihan Compatibility Variants to File
-const fs = require ('fs');
-const path = require ('path');
-const { app } = require ('electron').remote;
 const compatibilityVariants = require ('./lib/unicode/get-cjk-compatibility-variants.js');
 let sortedCompatibilityVariants = { };
 let sortedCharacters = Object.keys (compatibilityVariants).sort ((a, b) => a.codePointAt (0) - b.codePointAt (0));
@@ -10,6 +7,5 @@ for (let character of sortedCharacters)
      let variants = compatibilityVariants[character];
      sortedCompatibilityVariants[character] = (variants.length === 1) ? variants[0] : variants;
 }
-const jsonFile = path.join (app.getPath ('desktop'), 'unihan-compatibility-variants.json');
-fs.writeFileSync (jsonFile, $.stringify (sortedCompatibilityVariants, null, 4));
+let jsonFile = $.save ($.stringify (sortedCompatibilityVariants, null, 4), 'unihan-compatibility-variants.json');
 $.writeln (`Wrote Unihan compatibility variants to JSON file:\n${jsonFile}`);