+++ renderer/units/JavaScript Runner/samples/Unihan/Write Unihan Data to File.js
@@ -1,12 +1,8 @@
 // Write Unihan Data to File
-const fs = require ('fs');
-const path = require ('path');
-const { app } = require ('electron').remote;
 const { codePoints, fullSet } = require ('./lib/unicode/parsed-unihan-data.js');
 let start = window.performance.now ();
 let sortedCodePoints = { };
 fullSet.forEach (codePoint => { sortedCodePoints[codePoint] = codePoints[codePoint]; });
-const jsonFile = path.join (app.getPath ('desktop'), 'unihan-data.json');
-fs.writeFileSync (jsonFile, JSON.stringify (sortedCodePoints, null, 4));
+let jsonFile = $.save (JSON.stringify (sortedCodePoints, null, 4), 'unihan-data.json');
 let stop = window.performance.now ();
 $.writeln (`Wrote Unihan data to JSON file in ${((stop - start) / 1000).toFixed (2)} seconds:\n${jsonFile}`);