+++ renderer/units/JavaScript Runner/samples/Unihan/Write Unihan Data to Property Files.js
@@ -1,7 +1,4 @@
 // Write Unihan Data to Property Files
-const fs = require ('fs');
-const path = require ('path');
-const { app } = require ('electron').remote;
 const { codePoints, fullSet } = require ('./lib/unicode/parsed-unihan-data.js');
 let start = window.performance.now ();
 const properties = { };
@@ -17,18 +14,12 @@ for (let codePoint of fullSet)
         properties[property][codePoint] = dataProperties[property];
     }
 }
-const jsonDirectory = path.join (app.getPath ('desktop'), 'unihan-data');
-if (!fs.existsSync (jsonDirectory))
-{
-    fs.mkdirSync (jsonDirectory);
-}
 let jsonFiles = [ ];
 for (let property in properties)
 {
-    const jsonFile = path.join (jsonDirectory, `${property}.json`);
-    fs.writeFileSync (jsonFile, $.stringify (properties[property], null, 4));
+    let jsonFile = $.save ($.stringify (properties[property], null, 4), `${property}.json`, 'unihan-data');
     jsonFiles.push (jsonFile);
 }
 let stop = window.performance.now ();
-$.writeln (`Wrote Unihan data to JSON property files in ${((stop - start) / 1000).toFixed (2)} seconds:`);
+$.writeln (`Wrote Unihan data to ${jsonFiles.length} JSON property files in ${((stop - start) / 1000).toFixed (2)} seconds:`);
 jsonFiles.forEach (jsonFile => { $.writeln (jsonFile); });